api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=21.0.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.95,<1.0",
    "groq>=1.0.0",
    "httpx>=0.24,<1.0",
    "numpy>=1.24,<2.0",
    "pandas>=2.0,<3.0",
    "passlib[argon2]>=1.7.4",
    "prisma>=0.15.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1,<3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "requests>=2.28,<3.0",
    "scikit-learn>=1.2,<2.0",
    "urllib3>=1.26,<3.0",
    "uvicorn>=0.22,<1.0",
]
